image-mirror: copy active tags
The last updated date is a rough heuristic for identifying images that
are currently active. We now have an `activeTags` field on the repo that
we can use for this.

Continue to fetch images that were recently updated too. This captures
images that may not be active any more but were recently updated and we
may not have copied when the script was last ran.