python: move example from trainer-the-trainer to python (#74)
I really like this example and I like to use it outside of the context
of the trainer development course. So, I'm moving it under `python` and
linking back to it from the trainer development course README.

I've tweaked the presentation of it a little bit so it steps all the way
through from 'existing Dockerfile' to 'migrated Dockerfile'.