Clean up the trainer development workshops (#75)
* train-the-trainer: remove debugging demo in favour of debugging-distroless

We're using the same slides as the stand alone debugging workshop, so
let's use the same demo.

* train-the-trainer: tweak the README

* train-the-trainer: remove docker102

We're using the central Container 102 as part of the Trainer Development
course now. So we can remove this.

* train-the-trainer: tweak the README

* train-the-trainer: remove old java example

If we need it, we can retrieve it from the git history.

* keeping-up-with-latest: move tags-and-digests demo

I removed docker102 from the trainer development folder, but this
specific demo is still useful and I think it lives in the
keeping-up-with-latest workshop.

* train-the-trainer: remove duplicate demo-magic.sh